moved over all functions for the resetpassworddialogs
